Initial Commit
Webpage to change passwords for a given OU in active directory. This
only uses HTTP authentication and there is no SSL yet. This is not
recommended for users that must have secure accounts. It’s current
implementation is simply to allow our school librarians an easy to use
interface which gives them the ability to reset student passwords. The
users in the config.dat are users who will have access to the webpage.